+++ app/layout.tsx
@@ -1,7 +1,10 @@
 import { Metadata } from "next";
 import "@/styles/globals.css";
 import { ClerkProvider } from "@clerk/nextjs";
+import { Toaster } from "react-hot-toast";
 import Fathom from "@/components/Fathom";
+import Footer from "@/components/Footer";
+import Header from "@/components/Header";
 
 const title = "timetime.cc";
 const description = "Paste anything, get calendar events.";
@@ -35,7 +38,18 @@ export default function RootLayout({
       <html lang="en">
         <body>
           <Fathom />
-          {children}
+          <div className="mx-auto flex min-h-screen max-w-5xl flex-col items-center justify-center py-2">
+            <Toaster
+              position="top-center"
+              reverseOrder={false}
+              toastOptions={{ duration: 2000 }}
+            />
+            <Header />
+            <main className="mt-12 flex w-full flex-1 flex-col items-center justify-center px-4 sm:mt-20">
+              {children}
+            </main>
+            <Footer />
+          </div>
         </body>
       </html>
     </ClerkProvider>
